requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.5
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-crispy-forms==1.10.0
django-js-asset==1.2.2
django-taggit==1.3.0
gunicorn==20.0.4
Pillow==8.1.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2020.5
sqlparse==0.4.1
static3==0.7.0
Unipath==1.1
validate-email==1.3
whitenoise==5.2.0
